+++ src/StorageSync/StorageSync.Test/ScenarioTests/RegisteredServerTests.cs
@@ -65,6 +65,26 @@ namespace StorageSyncTests
         }
 
         /// <summary>
+        /// Defines the test method TestNewRegisteredServerWithIdentityError.
+        /// </summary>
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestNewRegisteredServerWithIdentityError()
+        {
+            TestRunner.RunTestScript("Test-NewRegisteredServerWithIdentityError"); 
+        }
+
+        /// <summary>
+        /// Defines the test method TestNewRegisteredServerWithIdentity.
+        /// </summary>
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestNewRegisteredServerWithIdentity()
+        {
+            TestRunner.RunTestScript("Test-NewRegisteredServerWithIdentity");
+        }
+
+        /// <summary>
         /// Defines the test method TestNewRegisteredServer.
         /// </summary>
         [Fact]