+++ src/StorageSync/StorageSync/StorageSyncService/SetStorageSyncServiceCommand.cs
@@ -123,8 +123,8 @@ namespace Microsoft.Azure.Commands.StorageSync.StorageSyncService
             Mandatory = false,
             HelpMessage = HelpMessages.StorageSyncServiceIdentityTypeParameter)]
         [ValidateSet(StorageSyncServiceIdentityType.systemAssigned,
-            StorageSyncServiceIdentityType.userAssigned,
-            StorageSyncServiceIdentityType.systemAssignedUserAssigned,
+            //StorageSyncServiceIdentityType.userAssigned,
+            //StorageSyncServiceIdentityType.systemAssignedUserAssigned,
             StorageSyncServiceIdentityType.none,
             IgnoreCase = true)]
         public string IdentityType { get; set; }